+++ OgreMain/include/OgreSceneManager.h
@@ -2581,7 +2581,7 @@ namespace Ogre {
             certain objects; see SceneQuery for details.
         */
         virtual AxisAlignedBoxSceneQuery* 
-            createAABBQuery(const AxisAlignedBox& box, unsigned long mask = 0xFFFFFFFF);
+            createAABBQuery(const AxisAlignedBox& box, uint32 mask = 0xFFFFFFFF);
         /** Creates a SphereSceneQuery for this scene manager. 
         @remarks
             This method creates a new instance of a query object for this scene manager, 
@@ -2595,7 +2595,7 @@ namespace Ogre {
             certain objects; see SceneQuery for details.
         */
         virtual SphereSceneQuery* 
-            createSphereQuery(const Sphere& sphere, unsigned long mask = 0xFFFFFFFF);
+            createSphereQuery(const Sphere& sphere, uint32 mask = 0xFFFFFFFF);
         /** Creates a PlaneBoundedVolumeListSceneQuery for this scene manager. 
         @remarks
         This method creates a new instance of a query object for this scene manager, 
@@ -2609,7 +2609,7 @@ namespace Ogre {
         certain objects; see SceneQuery for details.
         */
         virtual PlaneBoundedVolumeListSceneQuery* 
-            createPlaneBoundedVolumeQuery(const PlaneBoundedVolumeList& volumes, unsigned long mask = 0xFFFFFFFF);
+            createPlaneBoundedVolumeQuery(const PlaneBoundedVolumeList& volumes, uint32 mask = 0xFFFFFFFF);
 
 
         /** Creates a RaySceneQuery for this scene manager. 
@@ -2625,7 +2625,7 @@ namespace Ogre {
             certain objects; see SceneQuery for details.
         */
         virtual RaySceneQuery* 
-            createRayQuery(const Ray& ray, unsigned long mask = 0xFFFFFFFF);
+            createRayQuery(const Ray& ray, uint32 mask = 0xFFFFFFFF);
         //PyramidSceneQuery* createPyramidQuery(const Pyramid& p, unsigned long mask = 0xFFFFFFFF);
         /** Creates an IntersectionSceneQuery for this scene manager. 
         @remarks
@@ -2639,7 +2639,7 @@ namespace Ogre {
             certain objects; see SceneQuery for details.
         */
         virtual IntersectionSceneQuery* 
-            createIntersectionQuery(unsigned long mask = 0xFFFFFFFF);
+            createIntersectionQuery(uint32 mask = 0xFFFFFFFF);
 
         /** Destroys a scene query of any type. */
         virtual void destroyQuery(SceneQuery* query);
